Moving stream functions to stream.py
